tests/wpt/mozilla/tests/mozilla/microdata/extra_space_test.html

<!doctype html>
<meta charset="utf-8">
<html>
<head>
  <script src="../../../../web-platform-tests/resources/testharness.js"></script>
  <script src="../../../../web-platform/testharnessreport.js"></script>
  <title>Extra Space ItemProperty and ItemType test</title>
</head>

<body id="body" itemprop="       hello           world         " itemtype="       hello           world         ">

  <h1 id="h1" itemprop="       hello           world         " itemtype="       hello           world         ">Header H1</h1>
  <h2 id="h2" itemprop="       hello           world         " itemtype="       hello           world         ">Header H2</h2>

  <p id="p" itemprop="       hello           world         " itemtype="       hello           world         ">Paragraph</p>

  <form id="form" itemprop="       hello           world         " itemtype="       hello           world         ">
    <input id="input" itemprop="       hello           world         " itemtype="       hello           world         ">Input Field in form</input>
  </form>

  <ul id="ul" itemprop="       hello           world         " itemtype="       hello           world         ">
    <li id="li" itemprop="       hello           world         " itemtype="       hello           world         ">Unordered List Item</li>
  </ul>

  <meta id="meta" itemprop="       hello           world         " itemtype="       hello           world         ">Meta Tag</meta>

  <table id="table" itemprop="       hello           world         " itemtype="       hello           world         ">Table</table>

  <div id="div" itemprop="       hello           world         " itemtype="       hello           world         ">Hi</div>

  <a id="anchor" itemprop="       hello           world         " itemtype="       hello           world         "></a>

</body>

<script>
  test(function(){
    // Test body element
    var body = document.getElementById("body");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test h1 element
    var h1 = document.getElementById("h1");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test h2 element
    var h2 = document.getElementById("h2");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test p element
    var p = document.getElementById("p");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test input element
    var input = document.getElementById("input");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test form element
    var form = document.getElementById("form");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test li element
    var li = document.getElementById("li");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test ul element
    var ul = document.getElementById("ul");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test meta element
    var meta = document.getElementById("meta");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test table element
    var table = document.getElementById("table");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test div element
    var div = document.getElementById("div");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

    // Test anchor element
    var anchor = document.getElementById("anchor");
    assert_array_equals(body.propertyNames().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");
    assert_array_equals(body.itemtypes().sort(), ["hello", "world"].sort(), "The method should return 'hello world' ");

  });
</script>

</html>
